backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
pydantic==2.10.0
pydantic-settings==2.7.0
python-dotenv==1.0.1
httpx==0.28.0
mangum==0.19.0
scikit-learn==1.6.0
numpy==2.2.0
pandas==2.2.3
google-generativeai==0.8.0
celery[redis]==5.4.0
redis==5.2.0
alembic==1.14.0
python-multipart==0.0.18
